tests: Extract test common lib into a toplevel directory
