+++ test-common/eos-test-convenience.c
