+++ test-common/eos-test-convenience.h
