+++ test-common/misc-utils.c
