+++ tests/test-update-refspec-checkpoint.c
@@ -20,9 +20,9 @@
  *  - Sam Spilsbury <sam@endlessm.com>
  */
 
-#include "misc-utils.h"
-#include "spawn-utils.h"
-#include "eos-test-utils.h"
+#include <test-common/misc-utils.h>
+#include <test-common/spawn-utils.h>
+#include <test-common/eos-test-utils.h>
 
 #include <gio/gio.h>
 #include <locale.h>
